+++ examples/resources/cockroach_aws_private_endpoint_connection/resource.tf
@@ -4,6 +4,6 @@ variable "cluster_id" {
 }
 
 resource "cockroach_private_endpoint_connection" "cockroach" {
-  cluster_id     = var.cluster_id
-  endpoint_id    = "the endpoint id assigned by cloud provider to the client-side of the connection"
+  cluster_id  = var.cluster_id
+  endpoint_id = "the endpoint id assigned by cloud provider to the client-side of the connection"
 }