+++ internal/provider/person_user_data_source_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/cockroachdb/cockroach-cloud-sdk-go/pkg/client"
 	mock_client "github.com/cockroachdb/terraform-provider-cockroach/mock"
 	"github.com/golang/mock/gomock"
-	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 )
 
 func TestIntegrationPersonUserDataSource(t *testing.T) {
@@ -47,7 +47,7 @@ func TestIntegrationPersonUserDataSource(t *testing.T) {
 				Id:    "test-id",
 			},
 		}, nil, nil).
-		Times(5)
+		Times(3)
 
 	testPersonUserDataSource(t, email, true)
 }