+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Migrated the testing framework to
+  https://github.com/hashicorp/terraform-plugin-testing.
+
 ## [1.3.2] - 2024-03-13
 
 ### Changed